refactor: Moving files to db, workers and utils for later express implementation
